0406修正

+++ JS/backend.js
@@ -10,9 +10,10 @@ const token = {
 /* 取得後台訂單 (´・ω・`)人(´・ω・`)人(´・ω・`)人(´・ω・`)人(´・ω・`)*/
 function getOrders() {
   axios.get(url, token).then((res) => {
-    let orders = res.data.orders;
+    orders = res.data.orders;
     ordersRender(orders);
-  });
+    chartFilter(orders)
+  })
 }
 getOrders();
 
@@ -40,7 +41,8 @@ function ordersRender(data) {
       <td  class="order-date">${formatDate}</td>
       <td>${status}</td>
       <td>
-        <button class="del-order" data-set="${i.id}"> 移除 </button>
+        <button class="del-order" id="patchBTN" data-set="${i.id}" data-paid="${i.paid}"> 修改狀態 </button>
+        <button class="del-order" id="deleteOneBTN" data-set="${i.id}"> 移除 </button>  
       </td>
     </tr>
     `;
@@ -56,34 +58,55 @@ function timeformat(date) {
 }
 
 /* 刪除訂單 (´・ω・`)人(´・ω・`)人(´・ω・`)人(´・ω・`)人(´・ω・`)*/
-backendTable.addEventListener('click', delOrder);
-function delOrder(e) {
-  let id = e.target.dataset.set;
-  Swal.fire({
-    title: '確定要刪除訂單嗎？',
-    text: '刪掉不要後悔喔(´・ω・｀)',
-    icon: 'warning',
-    showCancelButton: true,
-    confirmButtonColor: '#a3cfee',
-    cancelButtonColor: '#f49f9f',
-    confirmButtonText: '刪',
-    cancelButtonText: '不要好了',
-  }).then((result) => {
-    if (result.isConfirmed) {
-      axios.delete(`${url}/${id}`, token).then((res) => {
-        Swal.fire({
-          position: 'center',
-          icon: 'success',
-          title: '成功刪除訂單',
-          showConfirmButton: false,
-          timer: 1500,
-        }).then(() => {
-          console.log(res);
-          getOrders();
+backendTable.addEventListener('click', delAndPatchOrder);
+function delAndPatchOrder(e) {
+  let id = e.target.dataset.set.set;
+  if( e.target.id == 'deleteOneBTN' ){
+    Swal.fire({
+      title: '確定要刪除訂單嗎？',
+      text: '刪掉不要後悔喔(´・ω・｀)',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#a3cfee',
+      cancelButtonColor: '#f49f9f',
+      confirmButtonText: '刪',
+      cancelButtonText: '不要好了',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete(`${url}/${id}`, token).then((res) => {
+          Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: '成功刪除訂單',
+            showConfirmButton: false,
+            timer: 1500,
+          }).then(() => {
+            console.log(res);
+            getOrders();
+          });
         });
-      });
+      }
+    });
+  } 
+  else {
+    let paid = e.target.dataset.paid ;
+    let putObj ={
+      data : {
+        id : id,
+        paid : !paid,
+      }
     }
-  });
+    const config = {
+      method: 'put',
+      url: url,
+      token,
+      data :putObj
+    }
+    axios(config)
+    .then( res => {
+      console.log(res)
+    })
+  }
 }
 
 /* 刪除全部訂單 (´・ω・`)人(´・ω・`)人(´・ω・`)人(´・ω・`)人(´・ω・`)*/
@@ -116,4 +139,43 @@ function delAllOrder() {
       });
     }
   })
+}
+
+
+function chartFilter(data) {
+	let editData = {};
+	data.forEach(item => {
+		if (editData[item.products[0].title] == undefined) {
+			editData[item.products[0].title] = 1;
+		} else {
+			editData[item.products[0].title] += 1;
+		}
+	});
+	let columns = [];
+	let product = Object.keys(editData);
+	product.forEach(item => {
+		columns.push([item, editData[item]]);
+	});
+	c3Chart(columns);
+}
+
+// c3 ☆.｡.:*・ﾟ ☆.｡.:*・ﾟ ☆.｡.:*・ﾟ ☆.｡.:*・ﾟ ☆.｡.:*・ﾟ ☆.｡.:*・ﾟ
+function c3Chart(columns) {
+	const chart = c3.generate({
+		bindto: "#chart",
+		data: {
+			columns: columns,
+			type: "donut",
+		},
+    color: {
+      pattern: ['#a4ceee', '#dfbff0', '#a3e4be', '#ebd6b6', '#b7e28f', '#98df8a']
+    },
+		donut: {
+			title: "銷售狀況 ( ^ω^ )",
+      width: 20,
+      label: {
+        show: false
+      }
+		}
+	});
 }
